refactor: 변경된 네이밍 import 재수정
